+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+- Mol* core dependency updated to 3.44.0
+- Removed Assembly Symmetry hack (now will hide assembly symmetry section for non-biological assemblies)
+
 ## [v3.1.3]
 - Added ``Assembly Symmetry`` to structure controls, requires setting ``symmetryAnnotation`` in initialization parameters
 - Keep sequence panel in settings even when initially hidden